Added a namespace containing physical constants. PeriodicTable is now a class with two methods to modify the mass and its variance of an element.

+++ src/constants/element.h
@@ -5,6 +5,6 @@ using namespace std;
 struct Element {
   string symbol;
   int atomicNumber;
-  float mass;
-  float massVariance;
+  double mass;
+  double massVariance;
 };
